+++ addon/index.ts
@@ -7,9 +7,8 @@ export {
   hasPendingWaiters,
 } from './waiter-manager';
 
-export { default as Token } from './token';
-export { default as AsyncTestWaiter } from './async-test-waiter';
 export { default as buildWaiter } from './build-waiter';
 export { default as waitForPromise } from './wait-for-promise';
+export { default as Token } from './token';
 
 export * from './types';
