--- addon/noop-test-waiter.ts
@@ -1,32 +1,0 @@
-import { TestWaiter, TestWaiterDebugInfo } from './types';
-
-import Token from './token';
-
-/**
- * A class providing a production, noop replacement for the {TestWaiter<T>} class.
- *
- * @public
- * @class TestWaiter<T>
- */
-export default class NoopTestWaiter implements TestWaiter {
-  name: string;
-  constructor(name: string) {
-    this.name = name;
-  }
-
-  beginAsync(): Token {
-    return this;
-  }
-
-  endAsync(): void {}
-
-  waitUntil(): boolean {
-    return true;
-  }
-
-  debugInfo(): TestWaiterDebugInfo[] {
-    return [];
-  }
-
-  reset(): void {}
-}
