+++ documentation.yml
@@ -7,10 +7,3 @@ toc:
       - getPendingWaiterState
       - hasPendingWaiters
       - _reset
-  - name: Async Test Waiter class
-    children:
-      - AsyncTestWaiter
-      - beginAsync
-      - endAsync
-      - waitUntil
-      - debugInfo
