+++ tests/unit/waiter-manager-noop-test.ts
@@ -1,7 +1,8 @@
+import { _reset, buildWaiter, getWaiters, register } from 'ember-test-waiters';
 import { module, test } from 'qunit';
+
 import { DEBUG } from '@glimmer/env';
-import { _reset, buildWaiter, register, getWaiters } from 'ember-test-waiters';
-import NoopTestWaiter from 'ember-test-waiters/noop-test-waiter';
+import NoopTestWaiter from 'ember-test-waiters/build-waiter';
 
 if (!DEBUG) {
   module('test-waiter | DEBUG: false', function(hooks) {
